# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/balances": {
      "get": {
        "operationId": "BalancesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Personal balances.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceListEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "List the current participant's nonzero balances by context and currency",
        "tags": ["balances"]
      }
    },
    "/api/v1/groups/{groupId}/balances": {
      "get": {
        "operationId": "GroupBalancesController_list",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
              "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group balances.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceListEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Get the current participant's balances in one group",
        "tags": ["groups", "balances"]
      }
    },
...
  },
  "info": {
    "title": "SPLITO API",
    "description": "Versioned API for SPLITO expense sharing and its financial journal",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Current SPLITO deployment"
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "SPLITO_SESSION"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token"
      }
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message", "requestId"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
                "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
              },
              "fieldErrors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["field", "message"],
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "BalanceLine": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "currency",
          "netAmountMinor",
          "owedAmountMinor",
          "receivableAmountMinor",
          "contextId",
          "contextName",
          "version"
        ],
        "properties": {
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "example": "INR"
          },
          "netAmountMinor": {
            "type": "string",
            "pattern": "^(?:0|-?[1-9][0-9]{0,18})$",
            "description": "Signed integer minor units serialized as a decimal string.",
            "example": "-1250"
          },
          "owedAmountMinor": {
            "type": "string",
            "pattern": "^(?:0|[1-9][0-9]{0,18})$",
            "description": "Integer minor units serialized as a decimal string.",
            "example": "1250"
          },
          "receivableAmountMinor": {
            "type": "string",
            "pattern": "^(?:0|[1-9][0-9]{0,18})$",
            "description": "Integer minor units serialized as a decimal string.",
            "example": "1250"
          },
          "contextId": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$",
            "example": "2f40d889-89f7-4fcb-99f0-8702f88d0c77"
          },
          "contextName": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "BalanceListEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data"],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BalanceLine"
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed input or identifier.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "A valid session is required.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated identity is not authorized.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource does not exist or is not visible to this identity.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Conflict": {
        "description": "The resource version, idempotency body, or financial state conflicts.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "PreconditionFailed": {
        "description": "The supplied resource version is stale.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "The uploaded file or multipart request exceeds its limit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "UnsupportedMediaType": {
        "description": "The upload is not an accepted image or multipart body.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "The financial-domain invariant was not satisfied.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "A required optimistic-concurrency header is missing.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "RateLimited": {
        "description": "The request rate limit was exceeded.",
        "headers": {
          "Retry-After": {
            "description": "Minimum seconds before retrying; 60 for cooldown or up to 900 for quota.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 900
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "InternalError": {
        "description": "An unexpected server error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "BadGateway": {
        "description": "The external SMS provider did not accept the invitation message.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "A required external delivery adapter is unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}